requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.42.0
datasets>=2.20.0
accelerate>=0.33.0
Pillow>=10.0.0
opencv-python>=4.9.0.80
gradio>=4.41.0
pytest>=8.2.0
types-Pillow>=10.2.0.20240822
types-requests>=2.32.0.20240914

